backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
greenlet>=3.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP client for LLM/Embedder
httpx==0.26.0
sse-starlette==2.0.0

# Vector database
qdrant-client>=1.7.1

# Document processing
python-docx==1.1.0
pypdf2==3.0.1
markdown==3.5.1

# Async file operations
aiofiles==23.2.1

# Utilities
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
